Add Business Details Questions for second form page
This adds the form structure and most of the questions for the Business
Details section of the new Additional Restrictions Grant form.

As there is one less page than before, it uses the BusinessDetails
component as that included most of the details required and I removed
the EligibilityCriteriaDetails component from the step to meet the
section labels.

Co-authored-by: Yusuf <yusuf@madetech.com>