Remove multiple grants functionality to handle single grant
This is a large change to change back from multiple grants to only handle a single grant amount awarded.
The main changes are in the naming or things and reducing down from three instances of grant types to one.
We also don't need to pass a variable to the `patchApplications` function which simplifies the update of the
application.