Attempt to set mime type on documents on upload
Currently when we upload documents, they are not storing the mime-type correctly, meaning that the document
is prompted to be downloaded to the user computer rather than opening in the browser as it should, to keep
on the right side of GDPR rules as we should know where any copies of PII (personally identifiable information)
are stored.

It's not clear if this will work as it can't be tested locally but in theory this shouldn't hurt and may solve the
problem. If this doesn't work, then we'll try and set the Content-Type header when a user id redirected to the
document in the application view component.