Update order of Application States
To make the statuses easier to use and make more sense with the similar
states closer to each other, this commit updates the order of the
statuses.

NOTE: The only currently used states on the production site are
"Declined for test" and "Unprocessed" which is why this change is
possible and they are the only states updated in `application_assessment`
in the migrations. It is not advised to change any other states in
other live applications where the states are already in use.