Attempt to set Content Type when opening documents
While we fixed the setting of mime Types when uploading files in a
previoud commit, this attempts to set the content type when
redirecting to open a file when an admin user makes the request.

Tests will be manual on staging, which is not ideal but local presigned
urls for the upload loacally don't work.