Use custom alphabet for nanoid id generator
Previously the `client_generated_id` could have any characters,
including `=`, `-` and other maths symbols. This was causing a problem
in exports as the data would be transformed by Excel in some cases to
a formula.

By using a cusom alphabet we can resolve this and also still be
confident collisions are going to be unlikely (checked using the nanoid
collision calculator at https://zelark.github.io/nano-id-cc/).